+++ apps/web/constants/push-subscriptions.ts
@@ -1,0 +1,2 @@
+// 계정 당 최대 구독 수 제한 (DoS 방지)
+export const MAX_SUBSCRIPTIONS_PER_USER = 10;
