--- apps/web/lib/push/index.ts
@@ -1,2 +1,0 @@
-export * from './config';
-export * from './web-push-client';
